# api/openapi.yaml
openapi: 3.0.3
info:
  title: AgnosTech API
  description: |
    API for AgnosTech - a tool to migrate cloud infrastructure to self-hosted Docker Compose stacks.

    ## Authentication
    Most endpoints require authentication via session cookie or Bearer token.
    Use the `/api/v1/auth/login` endpoint to obtain a session token.

    ## Error Responses
    All endpoints may return standard error responses:
    - `400 Bad Request` - Invalid request parameters
    - `401 Unauthorized` - Missing or invalid authentication
    - `404 Not Found` - Resource not found
    - `500 Internal Server Error` - Server error
    - `502 Bad Gateway` - External service connection failed
  version: 1.0.0
  contact:
    name: AgnosTech Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.homeport.local
    description: Local production server
...
paths:
...
  /api/v1/credentials/storage:
    post:
      tags: [Credentials]
      summary: Set storage credentials
      description: Store S3/MinIO credentials for the current session.
      operationId: setStorageCredentials
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StorageCredentials'
      responses:
        '200':
          description: Credentials stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: credentials stored
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Credentials]
      summary: Delete storage credentials
      description: Remove stored storage credentials for the current session.
      operationId: deleteStorageCredentials
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Credentials deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: credentials deleted
  /api/v1/credentials/database:
    post:
      tags: [Credentials]
      summary: Set database credentials
      description: Store PostgreSQL credentials for the current session.
      operationId: setDatabaseCredentials
      security:
        - sessionCookie: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatabaseCredentials'
      responses:
        '200':
          description: Credentials stored
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: credentials stored
        '400':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Credentials]
      summary: Delete database credentials
      description: Remove stored database credentials for the current session.
      operationId: deleteDatabaseCredentials
      security:
        - sessionCookie: []
        - bearerAuth: []
      responses:
        '200':
          description: Credentials deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: credentials deleted

  # Container Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: JWT Bearer token from login
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie from login
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
      required: [error]
...
    StorageCredentials:
      type: object
      required: [access_key, secret_key]
      properties:
        endpoint:
          type: string
          description: S3/MinIO endpoint (default localhost:9000)
        access_key:
          type: string
        secret_key:
          type: string
          format: password
    DatabaseCredentials:
      type: object
      required: [user, password]
      properties:
        host:
          type: string
          description: Database host (default localhost)
        port:
          type: integer
          minimum: 1
          maximum: 65535
          description: Database port (default 5432)
        database:
          type: string
          description: Database name (default postgres)
        user:
          type: string
        password:
          type: string
          format: password
        ssl_mode:
          type: string
          enum: [disable, require, verify-ca, verify-full]
          default: disable
...
security:
  - sessionCookie: []
  - bearerAuth: []